Adds tool folder, and moves renderCountdown
