+++ src/js/tools/renderCountdown.js
